Update jest for better transform support
