+++ packages/node/test/global-teardown.jest.ts
@@ -1,4 +1,5 @@
-module.exports = async function () {
+module.exports = async function() {
   console.log("tearing down");
+  // @ts-ignore
   global.ganacheServer.close();
-}
+};
